FIX - Telescope error when calling Git difftool
The latest version of Telescope has a delay of 10 ms when closing the
prompt buffer, which causes an error when calling 'Git difftool' that
switches focus to the diff buffer. This is fixed by manually closing the
prompt buffer before calling the diff command.

Fix also the title that appears on the Telescope results window when
searching for recent commits.